pyproject.toml
[project]
name = "test-multimodal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.25.0",
    "ruff>=0.12.3",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "yt-dlp>=2023.12.30",
    "pydantic>=2.0.0",
]
